tapdb: fix multiple issues around migration file handling

+++ tapdb/sqlc/migrations/000002_assets.up.sql
@@ -3,7 +3,7 @@
 -- information, along with indexing information is stored.
 -- TODO(roasbeef): also store SPV proof?
 CREATE TABLE IF NOT EXISTS chain_txns (
-    txn_id BIGINT PRIMARY KEY,
+    txn_id INTEGER PRIMARY KEY,
 
     txid BLOB UNIQUE NOT NULL,
 
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS chain_txns (
 -- outpoint itself, and also a references to the transaction that _spends_ that
 -- outpoint.
 CREATE TABLE IF NOT EXISTS genesis_points (
-    genesis_id BIGINT PRIMARY KEY,
+    genesis_id INTEGER PRIMARY KEY,
 
     -- TODO(roasbeef): just need the input index here instead?
     prev_out BLOB UNIQUE NOT NULL,
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS genesis_points (
 -- assets that we either created, or bootstrapped from the relevant Base
 -- Universe.
 CREATE TABLE IF NOT EXISTS assets_meta (
-    meta_id BIGINT PRIMARY KEY,
+    meta_id INTEGER PRIMARY KEY,
 
     meta_data_hash BLOB UNIQUE CHECK(length(meta_data_hash) = 32),
 
@@ -50,7 +50,7 @@ CREATE TABLE IF NOT EXISTS assets_meta (
 -- reference the genesis point which is also a necessary component for
 -- computing an asset ID.
 CREATE TABLE IF NOT EXISTS genesis_assets (
-    gen_asset_id BIGINT PRIMARY KEY,
+    gen_asset_id INTEGER PRIMARY KEY,
 
     asset_id BLOB UNIQUE,
 
@@ -72,7 +72,7 @@ CREATE INDEX IF NOT EXISTS asset_ids on genesis_assets(asset_id);
 -- full KeyLocator is stored so we can use these keys without actually storing
 -- the private keys on disk.
 CREATE TABLE IF NOT EXISTS internal_keys (
-    key_id BIGINT PRIMARY KEY,
+    key_id INTEGER PRIMARY KEY,
 
     -- We'll always store the full 33-byte key on disk, to make sure we're
     -- retaining full information.
@@ -88,7 +88,7 @@ CREATE TABLE IF NOT EXISTS internal_keys (
 -- tweaking the base group key by the associated genesis point. This table
 -- references the set of internal keys, and also the genesis_points table.
 CREATE TABLE IF NOT EXISTS asset_groups (
-    group_id BIGINT PRIMARY KEY,
+    group_id INTEGER PRIMARY KEY,
 
     tweaked_group_key BLOB UNIQUE NOT NULL CHECK(length(tweaked_group_key) = 33), 
 
@@ -107,7 +107,7 @@ CREATE TABLE IF NOT EXISTS asset_groups (
 -- asset ID must also be included. This table reference the asset ID it's used
 -- to create as well as the group key that signed the asset in the first place.
 CREATE TABLE IF NOT EXISTS asset_group_witnesses (
-    witness_id BIGINT PRIMARY KEY,
+    witness_id INTEGER PRIMARY KEY,
 
     -- The witness stack can contain either a single Schnorr signature for key
     -- spends of the tweaked group key, or a more complex script witness.
@@ -124,7 +124,7 @@ CREATE TABLE IF NOT EXISTS asset_group_witnesses (
 -- wallet, so the wallet is able to keep track of the amount of sats that are
 -- used to anchor Taproot assets.
 CREATE TABLE IF NOT EXISTS managed_utxos (
-    utxo_id BIGINT PRIMARY KEY,
+    utxo_id INTEGER PRIMARY KEY,
 
     outpoint BLOB UNIQUE NOT NULL,
 
@@ -163,7 +163,7 @@ CREATE TABLE IF NOT EXISTS managed_utxos (
 );
 
 CREATE TABLE IF NOT EXISTS script_keys (
-    script_key_id BIGINT PRIMARY KEY,
+    script_key_id INTEGER PRIMARY KEY,
 
     -- The actual internal key here that we hold the private key for. Applying
     -- the tweak to this gives us the tweaked_script_key.
@@ -184,7 +184,7 @@ CREATE TABLE IF NOT EXISTS script_keys (
 -- asset, along with the sibling taproot hash needed to properly reveal and
 -- spend the asset.
 CREATE TABLE IF NOT EXISTS assets (
-    asset_id BIGINT PRIMARY KEY,
+    asset_id INTEGER PRIMARY KEY,
     
     genesis_id BIGINT NOT NULL REFERENCES genesis_assets(gen_asset_id),
 
@@ -225,7 +225,7 @@ CREATE TABLE IF NOT EXISTS assets (
 -- asset. This then references the script key of an asset, creation a one to
 -- many relationship.
 CREATE TABLE IF NOT EXISTS asset_witnesses (
-    witness_id BIGINT PRIMARY KEY,
+    witness_id INTEGER PRIMARY KEY,
 
     asset_id BIGINT NOT NULL REFERENCES assets(asset_id) ON DELETE CASCADE,
 
@@ -243,7 +243,7 @@ CREATE TABLE IF NOT EXISTS asset_witnesses (
 );
 
 CREATE TABLE IF NOT EXISTS asset_proofs (
-    proof_id BIGINT PRIMARY KEY,
+    proof_id INTEGER PRIMARY KEY,
 
     -- We enforce that this value is unique so we can use an UPSERT to update a
     -- proof file that already exists.
@@ -260,7 +260,7 @@ CREATE TABLE IF NOT EXISTS asset_proofs (
 -- minting transaction which once signed and broadcast will actually create the
 -- assets.
 CREATE TABLE IF NOT EXISTS asset_minting_batches (
-    batch_id BIGINT PRIMARY KEY REFERENCES internal_keys(key_id),
+    batch_id INTEGER PRIMARY KEY REFERENCES internal_keys(key_id),
 
     -- TODO(roasbeef): make into proper enum table or use check to ensure
     -- proper values
@@ -281,7 +281,7 @@ CREATE INDEX IF NOT EXISTS batch_state_lookup on asset_minting_batches (batch_st
 -- asset_seedlings are budding assets: the contain the base asset information
 -- need to create an asset, but doesn't yet have a genesis point.
 CREATE TABLE IF NOT EXISTS asset_seedlings (
-    seedling_id BIGINT PRIMARY KEY,
+    seedling_id INTEGER PRIMARY KEY,
 
     -- TODO(roasbeef): data redundant w/ genesis_assets?
     -- move into asset details table?
